requirements.txt
docker-py==1.1.0
pyroute2==0.3.5
python-iptables==0.7.0
requests==2.6.0
six==1.9.0
